modules: led: Rework led module to remove PWM
Rework the led module to replace the PWM implementation with a GPIO
implementation. This is done as the nrf54l15 does not support PWM on all
leds, and the GPIO implementation is sufficient for most led use cases when
the leds are separate leds and not part of an RGB led.

Signed-off-by: Trond F. Christiansen <trond.christiansen@nordicsemi.no>